create javaLibs groovy script as a central point for different kinds of java pipelines
