+++ src/net/wooga/jenkins/pipeline/check/CheckCreator.groovy
@@ -1,5 +1,6 @@
 package net.wooga.jenkins.pipeline.check
 
+import net.wooga.jenkins.pipeline.check.steps.Step
 import net.wooga.jenkins.pipeline.config.Platform
 import net.wooga.jenkins.pipeline.config.UnityVersionPlatform
 
@@ -13,7 +14,7 @@ class CheckCreator {
     }
 
     Closure javaChecks(Platform platform, Step testStep, Step analysisStep) {
-        def mainClosure = createCheck(testStep, analysisStep).asRunnable(platform)
+        def mainClosure = createCheck(testStep, analysisStep).pack(platform)
         def catchClosure = {throw it}
         def finallyClosure = {
             jenkins.junit allowEmptyResults: true, testResults: "**/build/test-results/**/*.xml"
@@ -27,7 +28,7 @@ class CheckCreator {
     }
 
     Closure csWDKChecks(UnityVersionPlatform versionBuild, Step testStep, Step analysisStep) {
-        def mainClosure = createCheck(testStep, analysisStep).asRunnable(versionBuild.platform)
+        def mainClosure = createCheck(testStep, analysisStep).pack(versionBuild.platform)
         def catchClosure = { Throwable e ->
             if (versionBuild.optional) {
                 jenkins.unstable(message: "Unity build for optional version ${versionBuild.version} is found to be unstable\n${e.toString()}")
