+++ src/net/wooga/jenkins/pipeline/check/EnclosureCreator.groovy
@@ -1,8 +1,7 @@
 package net.wooga.jenkins.pipeline.check
 
+import net.wooga.jenkins.pipeline.check.steps.PackedStep
 import net.wooga.jenkins.pipeline.config.Platform
-
-import java.util.function.Consumer
 
 class EnclosureCreator {
 
@@ -14,7 +13,7 @@ class EnclosureCreator {
         this.buildNumber = buildNumber
     }
 
-    Closure withNodeAndEnv(Platform platform, Runnable mainCls, Closure catchCls, Runnable finallyCls) {
+    Closure withNodeAndEnv(Platform platform, PackedStep mainCls, Closure catchCls, PackedStep finallyCls) {
         def testEnvironment = platform.testEnvironment +
                 ["TRAVIS_JOB_NUMBER=${buildNumber}.${platform.name.toUpperCase()}"]
         return {
@@ -24,11 +23,11 @@ class EnclosureCreator {
             jenkins.node(nodeLabels) {
                 jenkins.withEnv(testEnvironment) {
                     try {
-                        mainCls.run()
+                        mainCls.call()
                     } catch (Exception e) {
                         catchCls?.call(e)
                     } finally {
-                        finallyCls?.run()
+                        finallyCls?.call()
                     }
                 }
             }
