+++ src/net/wooga/jenkins/pipeline/check/Enclosures.groovy
@@ -1,11 +1,9 @@
 package net.wooga.jenkins.pipeline.check
 
-
+import net.wooga.jenkins.pipeline.check.steps.PackedStep
 import net.wooga.jenkins.pipeline.config.DockerArgs
 import net.wooga.jenkins.pipeline.config.Platform
 import net.wooga.jenkins.pipeline.model.Docker
-
-import java.util.function.Consumer
 
 class Enclosures {
 
@@ -17,13 +15,13 @@ class Enclosures {
         this.enclosureCreator = enclosureCreator
     }
 
-    def withDocker(Platform platform, Runnable mainCls, Closure catchCls = {throw it}, Runnable finallyCls = {}) {
+    def withDocker(Platform platform, PackedStep mainCls, Closure catchCls = {throw it}, PackedStep finallyCls = {}) {
         enclosureCreator.withNodeAndEnv(platform, {
             docker.runOnImage(mainCls)
         }, catchCls, finallyCls)
     }
 
-    def simple(Platform platform, Runnable mainClosure,Closure catchCls = {throw it}, Runnable finallyCls = {}) {
+    def simple(Platform platform, PackedStep mainClosure, Closure catchCls = {throw it}, PackedStep finallyCls = {}) {
         enclosureCreator.withNodeAndEnv(platform, mainClosure, catchCls, finallyCls)
     }
 }
