+++ src/net/wooga/jenkins/pipeline/check/steps/PackedStep.groovy
@@ -1,0 +1,6 @@
+package net.wooga.jenkins.pipeline.check.steps
+
+//this is a fucking java.lang.Runnable, but muh jenkins sandbox
+interface PackedStep {
+    void call()
+}
