+++ src/net/wooga/jenkins/pipeline/check/steps/StepFunction.groovy
@@ -1,0 +1,7 @@
+package net.wooga.jenkins.pipeline.check.steps
+
+import net.wooga.jenkins.pipeline.config.Platform
+
+interface StepFunction {
+    void call(Platform platform)
+}
