+++ src/net/wooga/jenkins/pipeline/check/steps/StepWrapper.groovy
@@ -1,0 +1,7 @@
+package net.wooga.jenkins.pipeline.check.steps
+
+import net.wooga.jenkins.pipeline.config.Platform
+
+interface StepWrapper {
+    void call(Step step, Platform platform)
+}
